+++ lib/class/catalog.class.php
@@ -2581,9 +2581,9 @@ abstract class Catalog extends database_object
         debug_event(self::class, "import_playlist Parsed " . $playlist . ", found " . count($songs) . " songs", 5);
 
         if (count($songs)) {
-            $name        = $pinfo['extension'] . " - " . $pinfo['filename'];
+            $name = $pinfo['extension'] . " - " . $pinfo['filename'];
             // Search for existing playlist
-            $playlist_search = Playlist::get_playlists(true, null, $name);
+            $playlist_search = Playlist::get_playlists(null, $name);
             if (empty($playlist_search)) {
                 // New playlist
                 $playlist_id   = Playlist::create($name, 'public');